Refactor Lottie animation scaling using drawWithContent modifier
- Pull up aspect-ratio scaling and centering from canvas operations to
  top-level LottieAnimation via RemoteModifier.drawWithContent
- Remove the now-duplicated scaling from RenderShapes, so the scale is
  applied exactly once

Rendering is unchanged: all 33 Roborazzi goldens verify unmodified,
including the 16 LottieScalingDiffScreenshotTest aspect-ratio cases.
